Use dh_systemd_{enable,start} instead of dh_installsystemd
For the time being we want to support backports to Ubuntu 16.04
(Xenial), but its debhelper is too old for dh_installsystemd.

Systems with newer versions of debhelper (that do come with
dh_installsystemd) will continue to use dh_systemd_enable and
dh_systemd_start because debhelper compat is set to 10. We could add
some conditional logic to use dh_installsystemd instead of
dh_systemd_{enable,start} if available, but the added complexity is
probably not worthwhile.

I had originally set up overrides to use dh_installsystemd because I
misunderstood how dh works: I thought that the dh sequence was based
on the version of debhelper when it is actually based on the debhelper
compat level.