Upgrade to debhelper-compat 12
I had resisted upgrading from 10 because I wanted it to be easy to
backport the package to Ubuntu 16.04 (Xenial). Unfortunately, the
version of debhelper backported to Xenial does not support
dh_installinit --no-enable, which we require. Given that limitation,
and the numerous issues with debhelper < 12 involving dh_installinit,
dh_systemd_{enable,start}, and dh_installsystemd, I am giving up on
supporting Ubuntu 16.04.