debconf: Only prompt for PPP ip-up mode when use=if
The PPP ip-up hook does nothing unless the ddclient config contains a
`if=<interface>` setting with a matching interface, so there's no
point in prompting the user to answer ddclient/run_ipup if the user
wants to use a checkip service.

Also: Assume that the user wants daemon mode if the user does not want
ipup mode.